Separando codigo
